+++ pkgs/artisanal/lib/plugins.dart
@@ -36,4 +36,5 @@ export 'src/plugins/remote_surface_protocol.dart'
 export 'src/plugins/remote_surface_transport.dart'
     show RemotePluginJsonTransport;
 export 'src/plugins/remote_surface_channel.dart' show RemotePluginJsonChannel;
+export 'src/plugins/remote_surface_session.dart' show RemotePluginSession;
 export 'src/plugins/remote_surface_process.dart' show RemotePluginProcess;